update themeprovider files to prevent linting errors in ci tests
