+++ src/types/theme.ts
@@ -1,0 +1,13 @@
+export type Theme = "light" | "dark" | "system";
+
+export type ThemeProviderProps = {
+  children: React.ReactNode;
+  defaultTheme?: Theme;
+  storageKey?: string;
+};
+
+export type ThemeProviderState = {
+  theme: Theme;
+  effectiveTheme: "light" | "dark";
+  setTheme: (theme: Theme) => void;
+};
